requirements.txt
appdirs==1.4.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
Django==2.2.7
django-cors-headers==3.2.1
django-extensions==2.2.5
django-storages==1.8
djangorestframework==3.10.3
facebook-sdk==3.1.0
feedparser==5.2.1
ffmpy==0.2.3
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
idna==2.8
imageio==2.8.0
imageio-ffmpeg==0.4.1
isodate==0.6.0
link-preview==1.0
moviepy==1.0.2
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.34
pandas==0.25.3
Pillow==6.2.1
proglog==0.1.9
protobuf==3.11.1
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
PySocks==1.7.1
python-dateutil==2.8.1
python-ffmpeg-video-streaming==0.1.8
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tqdm==4.45.0
tweepy==3.8.0
uritools==3.0.0
urlextract==0.14.0
urllib3==1.25.7
youtube-dl==2020.3.24
